컴포넌트 분리, shallowEquals, deepEquals 함수작성
