+++ src/@lib/equalities/deepEquals.ts
@@ -1,3 +1,36 @@
 export function deepEquals<T>(objA: T, objB: T): boolean {
-  return objA === objB;
+  if (Object.is(objA, objB)) return true;
+
+  if (
+    typeof objA !== "object" ||
+    objA === null ||
+    typeof objB !== "object" ||
+    objB === null
+  ) {
+    return false;
+  }
+
+  if (Array.isArray(objA) && Array.isArray(objB)) {
+    if (objA.length !== objB.length) return false;
+    for (let i = 0; i < objA.length; i++) {
+      if (!deepEquals(objA[i], objB[i])) return false;
+    }
+    return true;
+  }
+  if (Array.isArray(objA) !== Array.isArray(objB)) return false;
+
+  const a = objA as Record<string, unknown>;
+  const b = objB as Record<string, unknown>;
+
+  const keysA = Object.keys(a);
+  const keysB = Object.keys(b);
+
+  if (keysA.length !== keysB.length) return false;
+
+  for (const key of keysA) {
+    if (!Object.prototype.hasOwnProperty.call(b, key)) return false;
+    if (!deepEquals(a[key], b[key])) return false;
+  }
+
+  return true;
 }
