+++ src/context/AppContext.tsx
@@ -1,0 +1,16 @@
+import { createContext, useContext } from "react";
+import { AppContextType } from "../types/types";
+// AppContext 타입 정의
+
+const AppContext = createContext<AppContextType | undefined>(undefined);
+
+// 커스텀 훅: useAppContext
+const useAppContext = () => {
+  const context = useContext(AppContext);
+  if (context === undefined) {
+    throw new Error("useAppContext must be used within an AppProvider");
+  }
+  return context;
+};
+
+export { AppContext, useAppContext };
